Fix misplaced and misspelled comments
The code below the comment in MovementSystem actually belonged to the previous comment. The code that comment is referring to seems to not exist. Later, when I know more, maybe I’ll delete the comment or implement what it describes.

In AttackSystem, I just added missing apostrophes to make the comment easier to read.