Fix EuclideanDistance by calculating sqrt
There would be a way to not have to calculate the square root, by checking whether to square the number right before comparing heuristic results, but that would make the code more complicated. Let’s see if the speed of sqrt is a problem in practice first.

Now all distances are Doubles.

I updated the tests and removed references to approximating Euclidean distance.